requirements.txt
aio_pika==9.4.0
fastapi==0.110.0
httpx==0.27.0
pydantic==2.6.3
pytest==8.1.1
uvicorn==0.28.0
python-dotenv==1.0.1